# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Bastion",
    "description": "ERP para pyme española. Contrato de la versión 1 de la API, bajo /api/v1/{modulo}/{recurso}. Se genera al compilar desde los controladores y los DTO: no se escribe a mano, y el cliente del frontal se genera de él.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/organizacion/impuestos/{id}": {
      "get": {
        "tags": [
          "Impuestos"
        ],
        "summary": "Devuelve un tramo.",
        "operationId": "Impuestos_Obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del tramo.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "put": {
        "tags": [
          "Impuestos"
        ],
        "summary": "Cambia el nombre y las cuentas contables de un tramo.",
        "operationId": "Impuestos_Modificar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del tramo.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarImpuestoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarImpuestoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarImpuestoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/v1/organizacion/impuestos/{id}/cierre": {
      "post": {
        "tags": [
          "Impuestos"
        ],
        "summary": "Pone fecha de fin a un tramo vigente.",
        "description": "Sub-recurso y no un campo del `PUT`: cerrar un tramo deja al código sin tipo a partir\ndel día siguiente, y eso va detrás de su propio permiso.",
        "operationId": "Impuestos_Cerrar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del tramo.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CerrarImpuestoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CerrarImpuestoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CerrarImpuestoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CerrarImpuestoDto": {
        "required": [
          "ultimoDia"
        ],
        "type": "object",
        "properties": {
          "ultimoDia": {
            "type": "string",
            "description": "Último día en que el tramo rige, incluido.",
            "format": "date"
          }
        },
        "description": "Lo que hace falta para cerrar un tramo vigente."
      },
...
      "ImpuestoDto": {
        "required": [
          "id",
          "codigo",
          "nombre",
          "tipo",
          "porcentaje",
          "vigenteDesde",
          "vigenteHasta",
          "cuentaRepercutido",
          "cuentaSoportado"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador del tramo.",
            "format": "uuid"
          },
          "codigo": {
            "type": "string",
            "description": "Código del impuesto, en mayúsculas. <b>Se repite entre tramos.</b>"
          },
          "nombre": {
            "type": "string",
            "description": "Nombre con el que se le conoce."
          },
          "tipo": {
            "type": "string",
            "description": "Clase de impuesto, como texto."
          },
          "porcentaje": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "El tipo aplicable, en tanto por ciento.",
            "format": "double"
          },
          "vigenteDesde": {
            "type": "string",
            "description": "Primer día en que rige, incluido.",
            "format": "date"
          },
          "vigenteHasta": {
            "type": [
              "null",
              "string"
            ],
            "description": "Último día en que rige, incluido, o nulo si sigue vigente.",
            "format": "date"
          },
          "cuentaRepercutido": {
            "type": [
              "null",
              "string"
            ],
            "description": "Cuenta contable del IVA repercutido, o nula."
          },
          "cuentaSoportado": {
            "type": [
              "null",
              "string"
            ],
            "description": "Cuenta contable del IVA soportado, o nula."
          }
        },
        "description": "Un tramo de un tipo impositivo, tal como sale de la API."
      },
...
      "ModificarImpuestoDto": {
        "required": [
          "nombre"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string",
            "description": "Nombre con el que se le conoce."
          },
          "cuentaRepercutido": {
            "maxLength": 9,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ],
            "description": "Cuenta contable del impuesto repercutido, o nula."
          },
          "cuentaSoportado": {
            "maxLength": 9,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ],
            "description": "Cuenta contable del impuesto soportado, o nula."
          }
        },
        "description": "Lo que se puede cambiar de un tramo ya abierto."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "testigoDeAcceso": {
        "type": "http",
        "description": "Testigo de acceso de quince minutos, en la cabecera Authorization. Lo emite POST /api/v1/identidad/sesiones y lo renueva POST .../sesiones/renovacion, que no lo lee de aquí sino de una cookie HttpOnly. El testigo de refresco no aparece en este contrato a propósito: el navegador no tiene que poder leerlo.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}